requirements.txt
asgiref==3.7.2
certifi==2022.12.05
dj-database-url==2.1.0
Django==3.2.23
djangorestframework==3.14.0
gunicorn==21.2.0
importlib-metadata==6.7.0
packaging==23.2
psycopg2==2.9.9
psycopg2-binary==2.9.6
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.7.1
whitenoise==6.4.0
wincertstore==0.2
zipp==3.15.0